Use quay.io/slopezpa/fedora-vgpu:latest
This image works, and using Fedora which has the required version of
mesa and recent packages for everything. Developing with old ancient
Ubuntu (22.04) is not useful.

Since we use Fedora for runtime, use it also for the build container,
based on podman gpu on macOS docs.

The docs show an example with Fedora 40, using special mesa version
that is not needed now. We use Fedora 42 to consume the latest versions.